Refactor Intelligence
