+++ server_core/__init__.py
@@ -8,6 +8,8 @@ from server_core.session_store import SessionStore
 from server_core.wordlist_store import WordlistStore
 from server_core.telemetry_collector import TelemetryCollector
 from server_core.modern_visual_engine import ModernVisualEngine
+from server_core.intelligence.cve_intelligence_manager import CVEIntelligenceManager
+from server_core.intelligence.intelligent_decision_engine import IntelligentDecisionEngine
 
 __all__ = [
     "ModernVisualEngine",
@@ -15,4 +17,6 @@ __all__ = [
     "TelemetryCollector",
     "SessionStore",
     "WordlistStore",
+    "CVEIntelligenceManager",
+    "IntelligentDecisionEngine"
 ]
